+++ CircleClickingGame/MainWindow.xaml.cs
@@ -165,8 +165,10 @@ namespace CircleClickingGame
 
         private void SettingsButton_Click(object sender, RoutedEventArgs e)
         {
-            SettingsWindow setwin = new SettingsWindow();
-            setwin.ShowDialog();
+            //SettingsWindow setwin = new SettingsWindow();
+            //setwin.ShowDialog();
+            testwindow t = new testwindow();
+            t.ShowDialog();
         }
 
     }
